Se completa la información de AnswerSession, y se muestran resultados

+++ Cuestionario/Cuestionario/Cuestionario.Model/AnswerSession.cs
@@ -20,9 +20,9 @@ namespace Questionnaire.Model
         public virtual string Username { get; set; }
 
         // TODO add comment that the duration is in seconds
-        public virtual int SessionDuration { get; set; }
+        public virtual double SessionDuration { get; set; }
 
-        public virtual int Score { get; set; }
+        public virtual double Score { get; set; }
 
         public virtual DateTime StartTime { get; set; }
